refactor(sidekick): split rust field annotations tests
